LP-3259: EOI Sign up button now a <button>
The sign up button on the "Get news and updates about the scheme" call to action on the homepage was a <a> link with no href attribute which was causing a failure on the Sortsite test.

It's now been changed to a button to be semantically correct and to remove the need for it to link anywhere.

I've also removed the onClick attribute and added a click listener to the site js. Note this uses optional chaining which is supported in all browsers in the standard GOV.UK support matrix, but isn't supported in IE11. If this is going to cause issue I can rewrite to a variable assignment and if condition.